In this committ, I started by shortening the code for getting the HTTP request from the client. I did this as I initially had the first line recv the data sent by the client and had it under another variable followed by the next line having that variable be equal to another variable so that it can be decoded later. So to fix this, I just changed the first line to the correct variable name 'message_bytes' and removed the second line as it was no longer necessary.
After completing this, I also included the code to send the client socket
all the contents of the cacheFile opened. Though before I did this, I joined
all the elements of the file into a single string and then encoded it to bytes
so that it can be sent to the client.