In this commit, I have adjusted the code to handle the cases when a cache file should not be created. So far I have done this for the following cases: - Statuscode 302 - Statuscode 404 - Statuscode 307 The above has been completed by wrapping the cache file creation code in an if statement that checks the status code of the response and if it includes any one of them or the cacheIt is not == True it would just skip the cache file creation.
I have also added a place holder variable for now called cacheIt
which will be used to determine if a cache file should be created
dependent on the cache directives (such as no-cache, no-store, private,
etc). Though the code for this still needs to be completed which would
be done in the next commit.

Furthermore, in this commit I currently have an error regarding how
the cache file is being created when the status code is 301.
This error has something do with the name of the cache file being
created based on the url. So I'll need to debug this further and
hopefully this will be fixed in the next commit.

Finally, regarding the previous commit, I forgot to mention why I made
the response section from the origin server the way I did.
Just a brief explanation of the code and why I did it,
the response variable was initialised as a byte string due to how
the sendall function only sends bytes. After this, I have a while
statement that will read the response from the origin server in chunks
and ammend it to the response variable. This keeps going until the next
response that is sent is empty and thus the if statement breaks out of
the loop. I did it this way to ensure that if the data is bigger than
the BUFFER_SIZE that it can still be obtained in the next go through.