In this commit I handled the case for the cache-control header max-age=<seconds>. I did this by first searching the cacheData for the cache-control header with the max-age directive. From there I had if statement that would run if max-age existed. Within this if statement, it would grab the age from the max-age header, grab the last time the cache was modified as well as the current time. I then subtracted the cacheTime from the currentTime and if it was less than the max age then the cache would still be fresh and thus would send the cache straight away from the proxy. Otherwise it would notice it and send a request to the origin server for a new cache.
After I completed this, I ran a bunch of tests on the proxy server
which most returned back successfull even though they took ages.
The only issue I ran into is with accessing a png.
This is due to how I am encoding and decoding the data which can't be
done with a png. This issue should hopefully be ironed out in the next
commit or at least stop the proxy from crashing when it fails.